interface: Don't mutate user options
This patch fixes cases where the 'opts' table passed by the user is
modified, making the external table unable to be reused in some
cases.

This was causing problems e.g. when dealing with properties, as the
user was having a predefined dbus options table, as the "interface"
key was changed internally to 'org.freedesktop.DBus.Properties'
without being changed back on return.